migrated to pydantic 2.0

+++ test.py
@@ -1,59 +1,8 @@
-from validators import validate_tag
 import json
 import re
 import copy
-from typing import Any
-
-with open("json_saves/autosave.json", "r") as f:
-    raw_json_save = json.load(f)
-
-json_save = copy.deepcopy(raw_json_save)
-
-# collect nation data and province in "nation_data" key
-json_save["nation_data"] = {}
-json_save["province_data"] = {}
-for k, v in raw_json_save.items():
-    if re.match(r"\b([A-Z]{3}|D[0-9]{2})\b", k):
-        json_save["nation_data"][k] = v
-        del json_save[k]
-    elif re.match(r"[0-9]{1,4}", k):
-        json_save["province_data"][k] = v
-        del json_save[k]
-
-
-def max_depth(d):
-    if isinstance(d, dict):
-        return 1 + (max(map(max_depth, d.values())) if d else 0)
-    return 0
-
-
-"""
-def collect_repeated_keys(dictionary: dict[str,Any]):
-    original = copy.deepcopy(dictionary)
-    all_keys = list(dictionary.keys())
-    already_collected = []
-    for key, value in original.items():
-        if f"{key}2" in all_keys:
-            #print(key)
-            collection = [value]
-            repeated_keys = []
-            i = 2
-            exhausted_keys = False
-            while not exhausted_keys:
-                try:
-                    key_to_find = f"{key}{i}"
-                    collection.append(dictionary[key_to_find])
-                    repeated_keys.append(key_to_find)
-                    i += 1
-                except KeyError:
-                    exhausted_keys = True
-            dictionary[key] = collection
-            for k in repeated_keys:
-                del dictionary[k]
-                all_keys.remove(k)
-            already_collected.append(key)
-    return dictionary
-"""
+import os
+from pydantic import ValidationError
 
 
 def collect_repeated_keys(dictionary):
@@ -100,9 +49,49 @@ def collect_repeated_keys(dictionary):
     return dictionary
 
 
-consolidated_json_save = collect_repeated_keys(json_save)
-json.dump(consolidated_json_save, open("consolidated_json_save.json", "w"))
+if not os.path.exists("consolidated_json_save.json"):
+    with open("json_saves/autosave.json", "r") as f:
+        raw_json_save = json.load(f)
+
+    json_save = copy.deepcopy(raw_json_save)
+
+    # collect nation data and province in "nation_data" key
+    json_save["nation_data"] = {}
+    json_save["province_data"] = {}
+    for k, v in raw_json_save.items():
+        if re.match(r"\b([A-Z]{3}|D[0-9]{2})\b", k):
+            json_save["nation_data"][k] = v
+            del json_save[k]
+        elif re.match(r"[0-9]{1,4}", k):
+            json_save["province_data"][k] = v
+            del json_save[k]
+    consolidated_json_save = collect_repeated_keys(json_save)
+    json.dump(consolidated_json_save, open("consolidated_json_save.json", "w"))
+else:
+    with open("consolidated_json_save.json", "r") as f:
+        json_save = json.load(f)
 
 from models import VicIISave
 
-VicIISave(**json_save)
+try:
+    VicIISave(**json_save)
+except ValidationError as e:
+    print(e)
+    """
+    error_string_lines = str(e).splitlines()
+    # just ignoring dict and list errors for repeated keys until the end
+    for n, line in enumerate(error_string_lines):
+        if "is not a valid list" in line or "is not a valid dict" in line:
+            pass
+        else:
+            try:
+                if (
+                    "is not a valid list" in error_string_lines[n + 1]
+                    or "is not a valid dict" in error_string_lines[n + 1]
+                ):
+                    pass
+                else:
+                    print(line)
+            except IndexError:
+                print(line)
+    """
